requirements.txt
azure-core==1.30.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.3
filelock==3.13.1
Flask==3.0.2
fsspec==2024.2.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
msal==1.26.0
msal-extensions==1.1.0
networkx==3.2.1
packaging==23.2
pillow==10.2.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.2.0
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
